CDAP-2780 move artifact validation to ArtifactRepository
Validation should occur for system artifacts that are scanned
from a directory as well as for artifacts added from the restful
apis. Therefore moving validation logic up a bit so that the same
code can be used to validate both.
